migrate FileButton tests
